@@ -1,12 +1,12 @@
 {
   "name": "imas_cg_cartoon",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "THE IDOLM@STER CINDERELLA GIRLS シンデレラガールズ劇場検索",
-  "main": "index.js",
+  "type": "module",
   "scripts": {
-    "start": "parcel src/index.html",
-    "build": "rm -rf dist/* && parcel build src/index.html --public-url .",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview"
   },
   "repository": {
     "type": "git",
@@ -18,15 +18,20 @@
     "url": "https://github.com/zaubermaerchen/imas_cg_cartoon_search/issues"
   },
   "homepage": "https://github.com/zaubermaerchen/imas_cg_cartoon_search#readme",
-  "devDependencies": {
-    "@vue/component-compiler-utils": "^2.6.0",
-    "parcel-bundler": "^1.12.4",
-    "typescript": "^3.7.4",
-    "vue-template-compiler": "^2.6.11"
-  },
   "dependencies": {
-    "element-ui": "^2.13.0",
-    "vue": "^2.6.11",
-    "vue-hot-reload-api": "^2.3.4"
+    "element-plus": "^2.2.15",
+    "vue": "^3.2.37"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.33.0",
+    "@typescript-eslint/parser": "^5.33.0",
+    "@vitejs/plugin-vue": "^3.0.3",
+    "@vue/eslint-config-typescript": "^11.0.0",
+    "eslint": "^8.22.0",
+    "eslint-plugin-vue": "^9.4.0",
+    "typescript": "^4.6.4",
+    "vite": "^3.0.7",
+    "vite-tsconfig-paths": "^3.5.0",
+    "vue-tsc": "^0.39.5"
   }
 }
