@@ -1,11 +1,7 @@
-import Vue from 'vue';
-import ElementUI from 'element-ui'
-import App from './app.vue';
-import 'element-ui/lib/theme-chalk/index.css'
+import { createApp } from 'vue'
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
+import App from '@/App.vue'
+import '@/index.css'
 
-Vue.use(ElementUI);
-
-new Vue({
-    el: "#app",
-    render: h => h(App)
-});
+createApp(App).use(ElementPlus).mount('#app')
